requirements.txt
pyspark
pandas
requests
psycopg2-binary
faker
